python: use PEP-420 compliant namespaces.
Currently there is no way to reference to a local protobuf module
on Linux if the dist-package python-protobuf is installed.
This is because the debian python-protobuf package creates an old-style
'google' module, and setuptools' pkg_resources does not seem to deal
properly with this mix.
According to pep-420 [1], the reccomended way of using namespaces today
is using pkgutil (instead of pkg_resources). This seems to deal properly
with the aforementioned problem.
More context in issue #1484.

[1] https://www.python.org/dev/peps/pep-0420/#namespace-packages-today